Split common module
