+++ crates/bitwarden-auth/src/token_management/mod.rs
@@ -1,9 +1,11 @@
 //! Token renewal module.
 
-mod common;
+mod middleware;
 mod password_manager_token_handler;
 #[cfg(feature = "secrets")]
 mod secrets_manager_token_handler;
+#[cfg(test)]
+pub(super) mod test_utils;
 
 pub use password_manager_token_handler::PasswordManagerTokenHandler;
 #[cfg(feature = "secrets")]
